chore: consolidate lib functions to be inside single files
