+++ src/lib/auth/sign-in.ts
@@ -7,7 +7,7 @@ import { resultAsyncToActionResult } from "@/types/error-typing";
 import { signInFormSchema } from "@/config/auth-schemas";
 import { createClient } from "@/utils/supabase/server";
 import { parseSchema } from "@/utils/parse-schema";
-import { getUserByAuthUuid } from "@/lib/users/query-uuid";
+import { getUserByAuthUuid } from "@/lib/users";
 import { completeSignUp } from "./complete-sign-up";
 
 type SignInError = {
