--- src/lib/campaigns/query-all.ts
@@ -1,9 +1,0 @@
-import { runQuery } from "@/utils/supabase-run";
-
-export const getCampaigns = () => 
-  runQuery((supabase) =>
-    supabase
-      .from("campaigns")
-      .select("*"),
-    "getCampaigns"
-  )
