--- src/lib/campaigns/query-shortened.ts
@@ -1,11 +1,0 @@
-import { runQuery } from "@/utils/supabase-run";
-
-export const getCampaign = ({ shortened }: { shortened: string }) =>
-  runQuery((supabase) =>
-    supabase
-      .from("campaigns")
-      .select("*, character_campaigns(*, characters(*))")
-      .eq("shortened", shortened)
-      .single(),
-    "getCampaign"
-  )
