+++ src/lib/characters/update.ts
@@ -6,10 +6,10 @@ import { z } from "zod";
 import { characterEditSchema } from "@/config/character-edit-schema";
 import { resultAsyncToActionResult } from "@/types/error-typing";
 import { parseSchema } from "@/utils/parse-schema";
-import { upsertCharacterRace } from "@/lib/races/upsert-character-race";
-import { deleteCharacterRace } from "@/lib/races/delete-character-race";
-import { insertClasses } from "@/lib/classes/insert";
-import { upsertRace } from "@/lib/races/upsert";
+import { upsertCharacterRace } from "@/lib/races";
+import { deleteCharacterRace } from "@/lib/races";
+import { insertClasses } from "@/lib/classes";
+import { upsertRace } from "@/lib/races";
 import { getPlayerUser } from "@/lib/player-user";
 import { deleteCharacterClass } from "./delete-character-class";
 import { upsertCharacterClass } from "./upsert-character-class";
