--- src/lib/dms/query-all.ts
@@ -1,7 +1,0 @@
-import { runQuery } from "@/utils/supabase-run";
-
-export const getDMs = () =>
-  runQuery((supabase) => supabase
-    .from("dms")
-    .select("*, users(*)")
-  )
