+++ src/lib/users.ts
@@ -1,0 +1,61 @@
+import { Tables } from "@/types/database.types";
+import { runQuery } from "@/utils/supabase-run";
+
+type User = Tables<"users">;
+type UserArgument = {
+  username: string;
+  knumber: string;
+  auth_user_uuid: string;
+};
+type GetUserByAuthUuidError = {
+  message: string;
+  code: "NOT_FOUND";
+};
+type GetUserByUsernameError = {
+  message: string;
+  code: "NOT_FOUND";
+}
+
+export const insertUser = (user: UserArgument) =>
+  runQuery<User>((supabase) => supabase
+    .from("users")
+    .insert(user)
+    .select("*")
+    .single()
+  );
+
+export const getUsers = () => 
+  runQuery((supabase) => supabase
+    .from("users")
+    .select("*, roles(*)")
+  );
+
+export const getUserByUsername = (username: string) =>
+  runQuery<User>((supabase) => supabase
+    .from("users")
+    .select("*")
+    .eq("username", username)
+    .single()
+  )
+  .mapErr((error) => error.message.includes("PGRST116")
+    ? {
+        message: `User not found: ${username}`,
+        code: "NOT_FOUND",
+      } as GetUserByUsernameError
+    : error
+  )
+
+export const getUserByAuthUuid = ({ authUserUuid }: { authUserUuid: string }) => 
+  runQuery((supabase) => supabase
+    .from("users")
+    .select("*")
+    .eq("auth_user_uuid", authUserUuid)
+    .single()
+  )
+  .mapErr((error) => error.message.includes("PGRST116")
+    ? {
+        message: `Public user with auth uuid '${authUserUuid}' not found from table users`,
+        code: "NOT_FOUND",
+      } as GetUserByAuthUuidError
+    : error
+  )
