--- src/lib/users/query-all.ts
@@ -1,7 +1,0 @@
-import { runQuery } from "@/utils/supabase-run";
-
-export const getUsers = () => 
-  runQuery((supabase) => supabase
-    .from("users")
-    .select("*, roles(*)")
-  )
